Update developerguide.md
Introduction:
Enterprise Connect (EC) is a micro service offered as part of the Predix environment.  It acts as a middleware to connect and transfer data between two discrete networks. Though the Enterprise Connect service is offered as part of Predix, it can connect any two independent networks. Therefore, it can even be used in a different environment such as Amazon Web Services (AWS).